pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dockermgr"
version = "0.4.1"
description = "A professional, menu-driven Docker management CLI with live dashboard and drill-down diagnostics."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Madrock Advisory"}]
dependencies = [
  "docker>=7.0.0",
  "psutil>=5.9.8",
  "rich>=13.7.1",
  "typer>=0.12.3",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0",
]

[project.scripts]
dockermgr = "dockermgr.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
